pyproject.toml
[project]
name = "lead-enrichment-automation"
version = "0.1.0"
description = "Enrich Salesforce Lead records with Google Maps data and automatically qualify or disqualify them using a configurable rule engine."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.14.1",
    "google-search-results>=2.4.2",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "rapidfuzz>=3.14.5",
    "requests",
    "simple-salesforce>=1.12.9",
    "tqdm>=4.67.3",
]
